requirements.txt
Flask==3.1.2
mysql-connector-python==9.6.0
python-dotenv==1.2.1
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.32.5
Werkzeug==3.1.3
